signup route refactored
I refactored the signup route. The signup process process is coomplete changed. I added a validation library called express-validator in order to validate email and password, I checked if the user with the given email exist, if yes we throw an error. Otherwise we create a new user, and also JsonWebtoken is created automatically as well is send back with id and email.